+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [8.3.0]
+
+### Added
+
+- Option to disable scroll to field on focus for A11y.
+
+### Fixed
+
+- Broken sorting for entries listing.
+
 ## [8.2.1]
 
 ### Fixed
@@ -1430,6 +1440,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[8.3.0]: https://github.com/infinum/eightshift-forms/compare/8.2.1...8.3.0
 [8.2.1]: https://github.com/infinum/eightshift-forms/compare/8.2.0...8.2.1
 [8.2.0]: https://github.com/infinum/eightshift-forms/compare/8.1.4...8.2.0
 [8.1.4]: https://github.com/infinum/eightshift-forms/compare/8.1.3...8.1.4